Organize methods
