+++ dwi_ml/gui/transformers_menus.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+
+def open_tto_window():
+    print("aallo tto")
+
+
+def open_ttst_window():
+    print("aallo ttst")
+
+
+def open_tto_from_checkpoint_window():
+    print("Allo TTO")
+
+
+def open_ttst_from_checkpoint_window():
+    print("Allo TTST")
+
