+++ dwi_ml/gui/utils/argparser_equivalent_for_gui.py
@@ -1,15 +1,5 @@
 # -*- coding: utf-8 -*-
-import os
-
 import dearpygui.dearpygui as dpg
-from dwi_ml.gui.utils.inputs import assert_single_choice_file_dialog
-
-from dwi_ml.gui.utils.my_styles import fixed_window_options, \
-    get_my_fonts_dictionary, set_global_theme
-from dwi_ml.gui.utils.window import callback_change_window
-from dwi_ml.io_utils import verify_checkpoint_exists, verify_which_model_in_path
-from dwi_ml.training.utils.batch_samplers import get_args_batch_sampler
-from dwi_ml.training.utils.experiment import get_mandatory_args_training_experiment
 
 style_input_item = {
     'indent': 600,
@@ -38,7 +28,7 @@ def manage_visual_default(default, dtype):
             raise ValueError("Data type {} not supported yet!".format(dtype))
 
 
-def _set_to_default_callback(sender, app_data, user_data):
+def set_to_default_callback(sender, app_data, user_data):
     default, dtype = user_data
     if app_data:  # i.e. == True
         nb_char = len('_default_checkbox')
@@ -46,25 +36,13 @@ def _set_to_default_callback(sender, app_data, user_data):
         dpg.set_value(dpg_item_name, manage_visual_default(default, dtype))
 
 
-def _log_value_remove_check(sender, app_data, user_data):
+def log_value_remove_check(sender, _, __):
     dpg.set_value(sender + '_default_checkbox', False)
-
-
-def file_dialog_ok_callback(sender, app_data):
-    assert_single_choice_file_dialog(app_data)
-    chosen_path = app_data['current_path']
-    if sender == "file_dialog_resume_from_checkpoint":
-        open_checkpoint_subwindow(chosen_path)
-    # else, nothing. We can access the value of the sender later.
-
-
-def file_dialog_cancel_callback(_, __):
-    pass
 
 
 def add_args_to_gui(args, section_name=None):
     if section_name is not None:
-        dpg.add_text('\nExperiment:')
+        dpg.add_text('\n' + section_name + ':')
 
     all_names = list(args.keys())
     all_mandatory = [n for n in all_names if n[0] != '-']
@@ -89,7 +67,7 @@ def _add_item_to_gui(arg_name, params, required):
 
     # User clicked on optional?
     if not required:
-        item_options = {'callback': _log_value_remove_check}
+        item_options = {'callback': log_value_remove_check}
     else:
         item_options = {}
 
@@ -119,7 +97,7 @@ def _add_item_to_gui(arg_name, params, required):
                                  default_value=True,
                                  tag=arg_name + '_default_checkbox',
                                  user_data=(default, dtype),
-                                 callback=_set_to_default_callback)
+                                 callback=set_to_default_callback)
 
     # 4. (Below: Help)
     dpg.add_text(params['help'], **style_help)
@@ -153,7 +131,7 @@ def _add_input_item_based_on_type(arg_name, params, item_options):
     return dtype
 
 
-def add_save_script_outfile():
+def add_save_script_outfilename_item():
     dpg.add_text('\n\n')
     with dpg.group(horizontal=True):
         dpg.add_text("Script output", indent=arg_name_indent)
@@ -161,101 +139,3 @@ def add_save_script_outfile():
                        tag='output_script', indent=300, width=600,
                        callback=lambda: dpg.show_item("file_dialog_output_script"))
     dpg.add_text("\n\n\n\n")
-
-
-def open_checkpoint_subwindow(chosen_path):
-    # toDo: if raises a FileNotFoundError, show a pop-up warning?
-    #  Currently prints the warning in terminal.
-    checkpoint_path = verify_checkpoint_exists(chosen_path)
-
-    model_dir = os.path.join(checkpoint_path, 'model')
-    model_type = verify_which_model_in_path(model_dir)
-    print("Model's class: {}".format(model_type))
-
-    if model_type == 'Learn2TrackModel':
-        open_l2t_from_checkpoint_window()
-    elif model_type == 'OriginalTransformerModel':
-        open_tto_from_checkpoint_window()
-    elif model_type == 'TransformerSrcAndTgtModel':
-        open_ttst_from_checkpoint_window()
-    else:
-        raise ValueError("This type of model is not managed by our DWIML' GUI.")
-
-
-def open_l2t_from_checkpoint_window():
-    with dpg.window(**fixed_window_options):
-        pass
-
-
-def open_tto_from_checkpoint_window():
-    print("Allo TTO")
-
-
-def open_ttst_from_checkpoint_window():
-    print("Allo TTST")
-
-
-def callback_ok_get_args(sender, app_data, args):
-    if sender == 'create_l2t_train_script':
-        all_values = {}
-        for arg_name in args.keys():
-            all_values[arg_name] = dpg.get_value(arg_name)
-
-        print("ALL VALUES: ")
-        print(all_values)
-        create_l2t_train_script(all_values)
-
-
-def prepare_and_show_train_l2t_window():
-    main_window = dpg.get_active_window()
-    dpg.hide_item(main_window)
-
-    if dpg.does_item_exist('train_l2t_window'):
-        dpg.set_primary_window('train_l2t_window', True)
-        dpg.show_item('train_l2t_window')
-    else:
-        # Create the Learn2track window.
-        my_fonts = get_my_fonts_dictionary()
-        with dpg.window(**fixed_window_options,
-                        tag='train_l2t_window') as l2t_window:
-            dpg.set_primary_window(l2t_window, True)
-
-            dpg.add_button(label='<-- Back', callback=callback_change_window,
-                           user_data=(l2t_window, main_window))
-
-            title = dpg.add_text('\nLEARN2TRACK:\n\n')
-            dpg.bind_font(my_fonts['default'])
-            dpg.bind_item_font(title, my_fonts['main_title'])  # NOT WORKING?
-
-            args = get_mandatory_args_training_experiment()
-            add_args_to_gui(args, 'Experiment')
-
-            dpg.add_text('\nBatch sampler:')
-            new_args = get_args_batch_sampler()
-            add_args_to_gui(new_args)
-            args.update(new_args)
-
-            add_save_script_outfile()
-            dpg.add_button(label='Create my script!', indent=1000,
-                           tag='create_l2t_train_script',
-                           callback=callback_ok_get_args,
-                           user_data=args, height=50)
-
-
-def create_l2t_train_script(all_values):
-    script = "l2t_train_model.py "
-    for arg_name, value in all_values.items():
-        if value is not None:
-            script += arg_name + ' ' + str(value)
-        elif not arg_name[0:2] == '--':
-            print("Some required values are not defined!")
-
-    print(script)
-
-
-def open_tto_window():
-    print("aallo tto")
-
-
-def open_ttst_window():
-    print("aallo ttst")
